requirements.txt
django==4.1.5
psycopg2-binary
pillow==9.4.0
django-storages==1.13.2
boto3==1.26.58
djangorestframework==3.14.0
django-cors-headers==3.13.0
